Added client/*, add db/services, removed lib token
This is not a working plugin yet, changes have been committed and
pushed so progress thus far can be pulled and worked on.

client/client.php and client/curl.php are used for testing the
webservice.

db/services was present in orvsd_installcourse and has been added with
modifications to orvsd_siteinfo.

lib.php has had it's token removed.