Initial siteinfo as webservice push
The webservice version of ORVSD_SiteInfo doesn't *work* yet, but it
doesn't break moodle when it is installed.

All changes made to the plugin in this push are taken from
ORVSD_InstallCourse and modified for SiteInfo. This should have
resulted in a working plugin and style consistency but right now
it does not have any working functionality.

Currently Moodle does not recognize that the webservice has any
functions.

The expected output is the output from a call to Moodle's 'site_info'
api. I may simplify it further to only output static text to make
sure it is a working skeleton before we attempt to add more
functionality.